Create UI of DictionaryDialog in Qt Designer
