+++ .gitignore
@@ -3,7 +3,7 @@ tools/
 profile/
 meta.json
 
-src/*/gui/forms/anki*
+src/*/gui/forms/
 src/*/gui/resources/anki*
 
 __pycache__/
