+++ src/spell_checker/dict.py
@@ -14,6 +14,7 @@ from aqt.utils import openFolder, showInfo, tooltip
 from .const import *
 from .bdicwriter import create_bdic
 from .anking_menu import get_anking_menu
+from .gui import dictionary_dialog
 
 
 def open_dict_dir() -> None:
@@ -47,7 +48,7 @@ class DictionaryManager:
 
         profile.setSpellCheckEnabled(False)
         profile.setSpellCheckLanguages({})
-        self._dicts = DictionaryDialog().getDictionaries()
+        self._dicts = dictionary_dialog.DictionaryDialog(parent=mw).getDictionaries()
 
         profile.setSpellCheckEnabled(profile.isSpellCheckEnabled())
         profile.setSpellCheckLanguages(self._dicts)
@@ -55,97 +56,6 @@ class DictionaryManager:
 
     def getDictionaries(self):
         return self._dicts
-
-
-class DictionaryDialog(QDialog):
-    def __init__(self):
-        QDialog.__init__(self)
-        self._setupDialog()
-        self._update()
-        self.exec()
-
-    def getDictionaries(self):
-        return self._dict
-
-    def _setupDialog(self):
-        self.setWindowTitle("Dictionaries")
-        self.setWindowModality(Qt.WindowModality.WindowModal)
-        self.resize(250, 250)
-
-        layout = QVBoxLayout()
-        self.list = QListWidget()
-        self.list.setAlternatingRowColors(True)
-        self.list.setSelectionMode(QAbstractItemView.SelectionMode.ExtendedSelection)
-        self.list.itemDoubleClicked.connect(self._toggle)
-
-        bws_btn = QPushButton("Browse")
-        bws_btn.clicked.connect(open_dict_dir)
-        custom_words_btn = QPushButton("Custom Dictionary")
-        custom_words_btn.clicked.connect(lambda _: CustomDicDialog(parent=self).exec())
-        en_btn = QPushButton("Enable")
-        en_btn.clicked.connect(self._enable)
-        dis_btn = QPushButton("Disable")
-        dis_btn.clicked.connect(self._disable)
-
-        control_box = QHBoxLayout()
-        control_box.addWidget(bws_btn)
-        control_box.addWidget(custom_words_btn)
-        control_box.addWidget(en_btn)
-        control_box.addWidget(dis_btn)
-
-        layout.addWidget(self.list)
-        layout.addLayout(control_box)
-        self.setLayout(layout)
-
-    def _update(self):
-        self._dict = []
-        self.list.clear()
-
-        try:
-            DICT_FILES = os.listdir(DICT_DIR)
-        except FileNotFoundError:
-            showInfo("Missing or no read/write permission to dictionary folder.")
-            return
-
-        for d in DICT_FILES:
-            if RE_DICT_EXT_ENABLED.search(d):
-                item = QListWidgetItem(d)
-                item.setData(Qt.ItemDataRole.UserRole, d)
-                self.list.addItem(item)
-                self._dict.append(d[:-5])
-
-        for d in DICT_FILES:
-            if RE_DICT_EXT_DISABLED.search(d):
-                item = QListWidgetItem(d)
-                item.setData(Qt.ItemDataRole.UserRole, d)
-                self.list.addItem(item)
-
-    def _enable(self):
-        sel = [i for i in range(self.list.count()) if self.list.item(i).isSelected()]
-        if sel:
-            for i in sel:
-                fn = self.list.item(i).text()
-                if RE_DICT_EXT_DISABLED.search(fn):
-                    f = os.path.join(DICT_DIR, fn)
-                    os.rename(f, f[:-9])
-        self._update()
-
-    def _disable(self):
-        sel = [i for i in range(self.list.count()) if self.list.item(i).isSelected()]
-        if sel:
-            for i in sel:
-                fn = self.list.item(i).text()
-                if RE_DICT_EXT_ENABLED.search(fn):
-                    f = os.path.join(DICT_DIR, fn)
-                    os.rename(f, f + ".disabled")
-        self._update()
-
-    def _toggle(self):
-        fn = self.list.currentItem().text()
-        if RE_DICT_EXT_ENABLED.search(fn):
-            self._disable()
-        else:
-            self._enable()
 
 
 class CustomDicDialog(QDialog):
